a public domain art asset for testing
http://opengameart.org/content/golden-ui

By Buch, CC0. Link to CC0 profile preferred:
http://opengameart.org/users/buch

Not sure I'll be using this in the actual
game, but it's an image for testing and
also a nice example of conventional UI
widgets and how to make art assets for them
(e.g. the 3x3 split).

It also looks good which can be quite a
motivator during dev.