jslint-ifying another file
It's somewhat sad to have to use --unparam,
this is a useful warning which would nice
to use only case-by-case exceptions for.
However, the syntax for exceptions is
very cumbersome.

It's better to have no warning than to
clutter the code with noise, or worse,
"workarounds" that add to the actual code.

The remaining jslint options seem to be more
or less Douglas Crockford's personal style
preferences, which I have no qualms about
disabling. More may come.