bugfix for smoothifier
Testing on android, this IS smoother than
before, so we'll keep it.

It occurs to me that the way to really
make this more performant is to only
redraw the parts of the screen that
actually need it. That's not something you
can do for every game (notably not
scrolling ones) but for the most part it
IS certainly something you can do for
this one.